Independización de funciones: rr_gen y dinamic_function, en archivos independientes
